Squashed commit of the following:
commit cf3e5e6a3cd41ce43772ae4368dd3b675cb41fa9
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Mon Dec 2 02:18:39 2024 +0200

    Finally.. Delete an announcement

commit ac3ccb353db3033c5cdfa1d640e9342d193eccb8
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Mon Dec 2 02:16:38 2024 +0200

    Same for edit announcements

commit 1b86c2475f62d6874637caca0bda2833be52a944
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Mon Dec 2 02:07:34 2024 +0200

    Use the helper function isCourseAdmin in the routes

commit 80ae4246fc23f54764b2a882d979ad4dc24dfdf1
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Mon Dec 2 02:03:34 2024 +0200

    create isCorseAdmin helper function

commit 527e96942e0cded40490f3f4514487f7cd2935b9
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Mon Dec 2 02:01:50 2024 +0200

    Use JWT in creating announcements

commit 6ba24cbb58f75fb8178ef43fee4f53241120a2a4
Author: Loay Al-Said <l.programming555@gmail.com>
Date:   Mon Dec 2 01:39:10 2024 +0200

    Use the JWT for fetching announcements

+++ test_server/routes/announcements.js
@@ -10,14 +10,16 @@ const {
   mockSections,
   repliesList,
 } = require('../mockData');
-const { getUserData } = require('../helperFunctions');
+const { getUserData, isCourseAdmin } = require('../helperFunctions');
 const db = require('../connect');
+const { verifyToken } = require('../middlewares/authMiddlewares');
+const { verify } = require('jsonwebtoken');
 
 
 const router = express.Router();
 
 // Get course announcements
-router.get('/courses/:id/announcements', (req, res) => {
+router.get('/courses/:id/announcements', verifyToken, (req, res) => {
   const courseId = req.params.id;
   const course = db.prepare('SELECT * FROM courses WHERE id = ?').get(courseId);
   if (!course) {
@@ -49,13 +51,17 @@ router.get('/courses/:id/announcements', (req, res) => {
 });
 
 // Create a course announcement
-router.post('/courses/:id/announcements', (req, res) => {
+router.post('/courses/:id/announcements', verifyToken, (req, res) => {
   const courseId = req.params.id;
-  const { userId, title, details } = req.body;
+  const { title, details } = req.body;
+  const userId = req.userId;
 
-  if (!userId || !title || !details) {
+  if (!title || !details) {
     return res.status(400).send({ message: 'Missing required fields' });
   }
+
+  
+  if (!isCourseAdmin(userId, courseId)) return res.status(403).send({ message: 'User is not a course admin' });
 
   try {
     const id = uuidv4();
@@ -88,15 +94,19 @@ router.post('/courses/:id/announcements', (req, res) => {
 });
 
 // Edit an announcement
-router.put('/announcements/:id', (req, res) => {
+router.put('/announcements/:id', verifyToken, (req, res) => {
   const { id } = req.params;
   const { title, details } = req.body;
+  const userId = req.userId;
 
   try {
-    const announcement = db.prepare('SELECT 1 FROM announcements WHERE id = ?').get(id);
-
-    if (!announcement) {
+    const { courseId } = db.prepare('SELECT courseId FROM announcements WHERE id = ?').get(id);
+    if (!courseId) {
       return res.status(404).send({ message: 'Announcement not found' });
+    }
+
+    if (!isCourseAdmin(userId, courseId)) {
+      return res.status(403).send({ message: 'User is not a course admin' });
     }
 
     db.prepare('UPDATE announcements SET title = ?, body = ? WHERE id = ?').run(title, details, id);
@@ -113,15 +123,17 @@ router.put('/announcements/:id', (req, res) => {
 });
 
 // Delete an announcement
-router.delete('/announcements/:id', (req, res) => {
+router.delete('/announcements/:id', verifyToken, (req, res) => {
   const announcementId = req.params.id;
   try {
     const announcement = db.prepare('SELECT * FROM announcements WHERE id = ?').get(announcementId);
-
     if (!announcement) {
       return res.status(404).send({ message: 'Announcement not found' });
     }
 
+    if (!isCourseAdmin(req.userId, announcement.courseId)) {
+      return res.status(403).send({ message: 'User is not a course admin' });
+    }
     db.prepare('DELETE FROM announcements WHERE id = ?').run(announcementId);
 
     res.status(200).json({ message: 'Announcement deleted successfully' });
